+++ CHANGELOG.md
@@ -11,10 +11,10 @@ All notable changes to this project will be documented in this file.
 ### Quality
 - Remove unused methods in `IexecHubService`. (#572)
 - Clean unused Replicate methods and update tests. (#573)
+- Clean unused `ReplicateStatus#RESULT_UPLOAD_REQUEST_FAILED`. (#575)
 ### Dependency Upgrades
-- Upgrade to `iexec-common` 8.0.0. (#571)
-- Add new `iexec-commons-poco` 1.0.2 dependency. (#571)
-- Update `iexec-commons-poco` to 1.0.2-NEXT-SNAPSHOT. (#574)
+- Upgrade to `iexec-common` 8.0.0-NEXT-SNAPSHOT. (#571 #575)
+- Add new `iexec-commons-poco` 1.0.2-NEXT-SNAPSHOT dependency. (#571 #574)
 
 ## [[8.0.1]](https://github.com/iExecBlockchainComputing/iexec-core/releases/tag/v8.0.1) 2023-03-20